fix(Test): fix fake member
