+++ test/api/user.register.test.js
@@ -1,38 +1,9 @@
 import request from 'supertest';
 import app from '../../src/app';
-import { renameSync, unlinkSync, writeFileSync } from 'fs';
 import { readLatestPassword } from '../util/mailReader';
 
-beforeAll(() => {
-    renameSync('./fakeMemberShip.json', './fakeMemberShip.backup.json');
-    renameSync('./mailWhitelist.json', './mailWhitelist.backup.json');
-
-    const membership = [
-        {
-            ID: 'ae4b61eb-92b7-4f87-9358-7a2cbbf03a81',
-            name: 'Gage',
-            email: 'Gage_Gislason@gmail.com'
-        }
-    ];
-
-    const memberJson = JSON.stringify(membership, null, 4);
-    writeFileSync('./fakeMemberShip.json', memberJson, 'utf8');
-
-    const mailWhiteList = [];
-
-    const mailWhiteListJson = JSON.stringify(mailWhiteList, null, 4);
-    writeFileSync('./mailWhitelist.json', mailWhiteListJson, 'utf8');
-
-});
-
-afterAll(() => {
-    unlinkSync('./fakeMemberShip.json');
-    unlinkSync('./mailWhitelist.json');
-    renameSync('./fakeMemberShip.backup.json', './fakeMemberShip.json');
-    renameSync('./mailWhitelist.backup.json', './mailWhitelist.json');
-});
-
 describe('Test user register', () => {
+
     test('It should block register request when no body is provided.', () => {
         return request(app)
             .post('/api/user/register')
